% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_debentures}
\alias{get_debentures}
\title{Imports the file Debentures.GLW}
\usage{
get_debentures()
}
\value{
A \code{tibble}.
}
\description{
Imports the file Debentures.GLW
}
\examples{
if (FALSE) {
  get_debentures()
}
}
