% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_ibovespa_stocks}
\alias{get_ibovespa_stocks}
\title{Imports the file cointegration_stocks_ibovespa.GLW}
\usage{
get_ibovespa_stocks()
}
\value{
A \code{tibble}.
}
\description{
Imports the file cointegration_stocks_ibovespa.GLW
}
\examples{
if (FALSE) {
  get_ibovespa_stocks()
}
}
