% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_ibrx100_stocks}
\alias{get_ibrx100_stocks}
\title{Imports the file cointegration_stocks_ibrx100.GLW}
\usage{
get_ibrx100_stocks()
}
\value{
A \code{tibble}.
}
\description{
Imports the file cointegration_stocks_ibrx100.GLW
}
\examples{
if (FALSE) {
  get_ibrx100_stocks()
}
}
