% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_indicadores_ibovespa}
\alias{get_indicadores_ibovespa}
\title{Imports the file Indicadores Ibovespa.GLW}
\usage{
get_indicadores_ibovespa()
}
\value{
A \code{tibble}.
}
\description{
Imports the file Indicadores Ibovespa.GLW
}
\examples{
if (FALSE) {
  get_indicadores_ibovespa()
}
}
