% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_risk_parity_brasil}
\alias{get_risk_parity_brasil}
\title{Imports the file risk_parity_brasil.GLW}
\usage{
get_risk_parity_brasil()
}
\value{
A \code{tibble}.
}
\description{
Imports the file risk_parity_brasil.GLW
}
\examples{
if (FALSE) {
  get_risk_parity_brasil()
}
}
