% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_screening_stocks}
\alias{get_screening_stocks}
\title{Imports the file Screening Ações.GLW}
\usage{
get_screening_stocks()
}
\value{
A \code{tibble}.
}
\description{
Imports the file Screening Ações.GLW
}
\examples{
if (FALSE) {
  get_screening_stocks()
}
}
