% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_tp}
\alias{get_tp}
\title{Imports the file Screening Titulos Publicos.GLW}
\usage{
get_tp()
}
\value{
A \code{tibble}.
}
\description{
Imports the file Screening Titulos Publicos.GLW
}
\examples{
if (FALSE) {
  get_tp()
}
}
