% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{get_uncovered_interest_rate_parity}
\alias{get_uncovered_interest_rate_parity}
\title{Imports the file PPP IUP.GLW}
\usage{
get_uncovered_interest_rate_parity()
}
\value{
A \code{tibble}.
}
\description{
Imports the file PPP IUP.GLW
}
\examples{
if (FALSE) {
  get_uncovered_interest_rate_parity()
}
}
